requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
pymysql>=1.0.3
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.28.0